# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: GTW IOT API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
...
paths:
...
  /user/resetUser:
    x-swagger-router-controller: userManage
    get:
      tags:
      - user
      security: []
      summary: resetUser in
      description: resetUser user
      # used as the method name of the controller
      operationId: resetUser
      parameters:
      - name: EmailAddress
        in: query
        description: EmailAddress to reset
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/UserModel"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"   
  /user/validateResetPasswordLink:
    x-swagger-router-controller: userManage
    get:
      tags:
      - user
      security: []
      summary: resetUser in
      description: resetUser user
      # used as the method name of the controller
      operationId: validateResetPasswordLink
      parameters:
      - name: EmailAddress
        in: query
        description: EmailAddress to reset
        required: true
        type: string
      - name: VerificationCode
        in: query
        description: VerificationCode to reset
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/UserModel"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"   
  /user/updatePassword:
    x-swagger-router-controller: userManage
    get:
      tags:
      - user
      security: []
      summary: updatePassword in
      description: updatePassword user
      # used as the method name of the controller
      operationId: updatePassword
      parameters:
      - name: EmailAddress
        in: query
        description: EmailAddress to reset
        required: true
        type: string
      - name: Password
        in: query
        description: VerificationCode to reset
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/UserModel"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"   
  /user/createUser:
    x-swagger-router-controller: userManage
    post:
      tags:
      - user
      security: []
      summary: create a new user
      description: create a new user
      # used as the method name of the controller
      operationId: createUser
      parameters:
      - in: body
        name: body
        description: User object thats need to be added
        required: true
        schema:
          $ref: '#/definitions/UserModel'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/UserModel"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/updateUser:
    x-swagger-router-controller: userManage
    put:
      tags:
      - user
      summary: update a user
      description: update a user
      # used as the method name of the controller
      operationId: updateUser
      parameters:
      - in: body
        name: body
        description: User object thats need to be updated
        required: true
        schema:
          $ref: '#/definitions/UserModel'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/deleteUser:
    x-swagger-router-controller: userManage
    delete:
      tags:
      - user
      summary: delete a user
      description: delete a user
      # used as the method name of the controller
      operationId: deleteUser
      parameters:
      - name: userID
        in: query
        description: User object thats need to be deleted
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/getSettings:
    x-swagger-router-controller: userManage
    get:
      tags:
      - user
      summary: get all user settings
      description: get all user gettings
      # used as the method name of the controller
      operationId: getSettings
      parameters:
      - name: userID
        in: query
        description: get all user settings
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/UserSettings"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/updateSettings:
    x-swagger-router-controller: userManage
    post:
      tags:
      - user
      summary: update all user settings
      description: update all user gettings
      # used as the method name of the controller
      operationId: updateSettings
      parameters:
      - in: body
        name: body
        description: User object thats need to be updated
        required: true
        schema:
          $ref: '#/definitions/UserModel'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/activate:
    x-swagger-router-controller: userManage
    get:
      tags:
      - user
      security: []
      summary: activate user
      description: activate user
      # used as the method name of the controller
      operationId: activate
      parameters:
      - name: UserID
        in: query
        description: User ID
        required: true
        type: string
      - name: ActivateKey
        in: query
        description: ActivateKey
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/getUserIDbyApiKey:
    x-swagger-router-controller: userManage
    get:
      tags:
      - user
      description: get the UserID using the Api-Key
      operationId: getUserbyApiKey
      parameters:
      - name: ApiKey
        description: bf
        required: true
        in: query
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/UserModel"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  UserModel:
    type: object
    properties:
      UserID:
        type: string
      Active:
        type: number
      EmailAddress:
        type: string
      Password:
        type: string
...
  UserSettings:
    type: object
    properties:
      SettingName:
        type: string
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string